+++ CHANGELOG.md
@@ -5,9 +5,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [0.3.1] - 2020-12-26 🎄
-- Implements an abstraction layer to handle CORS
+## [0.3.1] - 2020-12-27 🎄
+- Implements an abstraction layer to [handle CORS](https://www.neoteroi.dev/blacksheep/cors/)
 - Improves the code API to handle [response cookies](https://www.neoteroi.dev/blacksheep/responses/#setting-cookies)
+- Improves the default handling of authorization for request handlers (#69)
 - Adds more binders: `FromText`, `FromBytes`, `RequestMethod`, `RequestURL`
 - Improves `FromJson` binder to support returning the dictionary after JSON deserialization
 - Improves the default bad request response for invalid dataclass